requirements.txt
# Apache Airflow and dependencies
apache-airflow==2.8.1
apache-airflow-providers-common-sql==1.10.0

# Data processing
pandas==2.1.4
numpy==1.26.3

# Database
sqlite3

# Utilities
python-dateutil==2.8.2